refactor(auth): refactor jsp pages, home and base servlet
change links in jsp pages
add method for handling with previous page path to the base servlet
add code to handle redirecting to the previous page to the auth filter
add second path to the home servlet
