+++ src/main/webapp/sign-in.jsp
@@ -28,7 +28,7 @@
                 </p>
                 <button type="submit" class="btn btn-primary btn-block" style="background-color: black; border-color: black">Sign In</button>
             </form>
-            <p class="mt-3 text-center">Don't have an account? <a href="sign-up.jsp">Sign Up</a></p>
+            <p class="mt-3 text-center">Don't have an account? <a href="sign-up">Sign Up</a></p>
         </div>
     </div>
 </div>
